Port across rest of the Bedrock codecs

+++ bedrock/bedrock-codec/src/main/java/org/cloudburstmc/protocol/bedrock/data/command/CommandData.java
@@ -42,8 +42,8 @@ public class CommandData {
     public static class Builder {
         private final String name;
         private final String description;
-        private final byte flags;
-        private final byte permission;
+        private final int flags;
+        private final int permission;
         private final int aliases;
         private CommandParamData.Builder[][] overloads;
     }
